requirements.txt
# HKD 프로젝트 Python 의존성
# pip install -r requirements.txt

# HTTP 및 SSL
requests>=2.31.0
certifi>=2024.2.2

# HTML 파싱
beautifulsoup4>=4.12.0
lxml>=5.1.0

# 테스트
pytest>=8.0.0
pytest-cov>=4.1.0
responses>=0.25.0

# 유틸리티
python-dotenv>=1.0.0
